# speed up `collector()` helper