This helper takes up most all of the time in `collect_predictions()`. tune will see some speedup here, but `stacks::add_candidates()` especially benefits.

With `main` dev:

``` r
library(tidymodels)

res <- 
  fit_resamples(
    linear_reg(), 
    mpg ~ ., 
    bootstraps(mtcars),
    control = control_grid(save_pred = TRUE)
  )

bench::mark(
  predictions = tune:::collector(res, ".predictions"),
  metrics = tune:::collector(res, ".metrics"),
  check = FALSE
)
#> # A tibble: 2 × 6
#>   expression       min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>  <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 predictions   3.12ms    3.4ms      283.   621.4KB     6.24
#> 2 metrics       3.02ms   3.35ms      298.    36.5KB     6.22
```

With this PR:

``` r
#> # A tibble: 2 × 6
#>   expression       min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>  <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 predictions    433µs    445µs     2236.    75.7KB     6.18
#> 2 metrics        328µs    337µs     2960.    12.8KB     8.26
```